+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 - Fix some bugs around move-link navigation throughout the UI ([#647](https://github.com/ben/foundry-ironsworn/pull/647))
+- Add a tour for the Starforged character sheet ([#648](https://github.com/ben/foundry-ironsworn/pull/648))
 
 ## 1.20.21